# packages/arkeon/spec/openapi.snapshot.json
{
  "components": {
...
    "schemas": {
      "Actor": {
        "properties": {
          "can_publish_public": {
            "type": "boolean"
          },
          "created_at": {
            "example": "2026-01-01T00:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV",
            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$/i",
            "type": "string"
          },
          "is_admin": {
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "agent"
            ],
            "type": "string"
          },
          "max_read_level": {
            "description": "0=PUBLIC, 1=INTERNAL, 2=TEAM, 3=CONFIDENTIAL, 4=RESTRICTED",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "max_write_level": {
            "description": "0=PUBLIC, 1=INTERNAL, 2=TEAM, 3=CONFIDENTIAL, 4=RESTRICTED",
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "owner_id": {
            "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV",
            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$/i",
            "type": [
              "string",
              "null"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "status": {
            "enum": [
              "active",
              "suspended",
              "deactivated"
            ],
            "type": "string"
          },
          "updated_at": {
            "example": "2026-01-01T00:00:00.000Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "max_read_level",
          "max_write_level",
          "is_admin",
          "can_publish_public",
          "owner_id",
          "properties",
          "status",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "example": "invalid_body",
                "type": "string"
              },
              "details": {
                "$ref": "#/components/schemas/JsonObject"
              },
              "message": {
                "example": "Invalid request",
                "type": "string"
              },
              "request_id": {
                "example": "req_123",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JsonObject": {
        "additionalProperties": {},
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Arkeon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/actors": {
      "get": {
        "operationId": "listActors",
        "parameters": [
          {
            "description": "Max results (default 50, max 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "Max results (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Pagination cursor",
              "type": "string"
            }
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "description": "Filter by status",
              "enum": [
                "active",
                "suspended",
                "deactivated"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by kind",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "description": "Filter by kind",
              "enum": [
                "agent"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actors": {
                      "items": {
                        "$ref": "#/components/schemas/Actor"
                      },
                      "type": "array"
                    },
                    "cursor": {
                      "description": "Opaque pagination cursor",
                      "example": "eyJ0Ijoi...",
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "actors",
                    "cursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Actor listing"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          }
        },
        "summary": "List actors (paginated)",
        "tags": [
          "Actors"
        ],
        "x-arke-auth": "required",
        "x-arke-related": [
          "POST /actors",
          "GET /actors/{id}"
        ],
        "x-arke-rules": []
      },
      "post": {
        "operationId": "createActor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "kind": {
                    "description": "Actor kind (only 'agent' is supported)",
                    "enum": [
                      "agent"
                    ],
                    "type": "string"
                  },
                  "max_read_level": {
                    "description": "Max read level (0-4, default 1=INTERNAL)",
                    "maximum": 4,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "max_write_level": {
                    "description": "Max write level (0-4, default 1=INTERNAL)",
                    "maximum": 4,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "properties": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/JsonObject"
                      },
                      {
                        "description": "Actor properties"
                      }
                    ]
                  }
                },
                "required": [
                  "kind"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actor": {
                      "$ref": "#/components/schemas/Actor"
                    },
                    "api_key": {
                      "description": "Plaintext API key (only shown once)",
                      "type": "string"
                    }
                  },
                  "required": [
                    "actor",
                    "api_key"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Actor created with initial API key"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "Create a new actor and its initial API key",
        "tags": [
          "Actors"
        ],
        "x-arke-auth": "required",
        "x-arke-related": [
          "GET /actors/{id}",
          "GET /actors"
        ],
        "x-arke-rules": [
          "Cannot grant max_read_level higher than your own",
          "Cannot grant max_write_level higher than your own"
        ]
      }
    },
...
    "/resolve": {
      "post": {
        "operationId": "resolveSubject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "candidate_filter": {
                    "description": "Additional Meilisearch filters, e.g. [\"type = \\\"wiki\\\"\"]",
                    "items": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "maxItems": 10,
                    "type": "array"
                  },
                  "context": {
                    "description": "Surrounding prose passed to the LLM judge as disambiguation hints — never used as a search query",
                    "maxLength": 4000,
                    "type": "string"
                  },
                  "description": {
                    "description": "Free-form description — used both for query generation and as disambiguation context",
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "keywords": {
                    "description": "Alternate names/search phrasings (each becomes its own Meilisearch query)",
                    "items": {
                      "maxLength": 200,
                      "minLength": 1,
                      "type": "string"
                    },
                    "maxItems": 20,
                    "type": "array"
                  },
                  "label": {
                    "description": "Canonical name of the thing to match",
                    "maxLength": 400,
                    "minLength": 1,
                    "type": "string"
                  },
                  "limit": {
                    "description": "Max matches to return (default: all)",
                    "maximum": 20,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "llm_step": {
                    "description": "Which LLM step config to use (defaults to \"resolve\")",
                    "enum": [
                      "resolve",
                      "exists",
                      "dedup"
                    ],
                    "type": "string"
                  },
                  "space_id": {
                    "description": "If set, restrict candidates to this space",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$/i",
                    "type": "string"
                  }
                },
                "required": [
                  "label"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actor_read_level": {
                      "description": "The read_level the candidate pool was filtered against",
                      "maximum": 4,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "matches": {
                      "items": {
                        "properties": {
                          "confidence": {
                            "description": "1.0 = exact normalized-label match; 0.8 = LLM-confirmed; 0 = no match",
                            "type": "number"
                          },
                          "id": {
                            "description": "ULID string",
                            "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$/i",
                            "type": "string"
                          },
                          "rationale": {
                            "description": "Short explanation of the match decision",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "confidence"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "matches",
                    "actor_read_level"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Matches (possibly empty) in rank order"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Find existing entities matching a subject via multi-query Meilisearch + LLM judge",
        "tags": [
          "Resolve"
        ],
        "x-arke-auth": "required",
        "x-arke-related": [
          "POST /wiki",
          "GET /search"
        ],
        "x-arke-rules": [
          "Candidate pool is filtered to entities visible under the caller's max_read_level",
          "Returns 503 if no LLM provider is configured (see ~/.arkeon/llm.json)"
        ]
      }
    },
    "/search": {
      "get": {
        "description": "Keyword search with typo tolerance, prefix matching, and relevance ranking. Results are automatically filtered by the caller's read clearance level. Use the read_level parameter to restrict results below your clearance. Use view=expanded to include each result's relationships with counterpart summaries.",
        "operationId": "searchEntities",
        "parameters": [
          {
            "description": "Projection: summary | expanded. Default returns all fields. expanded adds _relationships.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "description": "Projection: summary | expanded. Default returns all fields. expanded adds _relationships.",
              "enum": [
                "summary",
                "expanded"
              ],
              "type": "string"
            }
          },
          {
            "description": "Comma-separated field list",
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "description": "Comma-separated field list",
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Search query string",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Search query string",
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Filter by entity type",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "description": "Filter by entity type",
              "type": "string"
            }
          },
          {
            "description": "Filter by kind (entity or relationship). Defaults to excluding relationships.",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "description": "Filter by kind (entity or relationship). Defaults to excluding relationships.",
              "type": "string"
            }
          },
          {
            "description": "Scope search to a space ULID",
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "description": "Scope search to a space ULID",
              "type": "string"
            }
          },
          {
            "description": "Restrict results to this read level or below (cannot exceed your clearance)",
            "in": "query",
            "name": "read_level",
            "required": false,
            "schema": {
              "description": "Restrict results to this read level or below (cannot exceed your clearance)",
              "maximum": 4,
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Page size (default 50, max 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "Page size (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Offset for pagination (default 0)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "description": "Offset for pagination (default 0)",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Max relationships per result when view=expanded (default 5, max 100)",
            "in": "query",
            "name": "rel_limit",
            "required": false,
            "schema": {
              "description": "Max relationships per result when view=expanded (default 5, max 100)",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "estimatedTotalHits": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    },
                    "results": {
                      "items": {
                        "anyOf": [
                          {
                            "$ref": "#/components/schemas/Entity"
                          },
                          {
                            "$ref": "#/components/schemas/ExpandedEntity"
                          }
                        ]
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "results",
                    "estimatedTotalHits",
                    "limit",
                    "offset"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Search results ordered by relevance. When view=expanded, each result includes _relationships and _relationships_truncated."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Full-text search across entities via Meilisearch",
        "tags": [
          "Search"
        ],
        "x-arke-auth": "optional",
        "x-arke-rules": [
          "Results filtered by your classification clearance"
        ]
      }
    },
...
    "/spaces": {
      "get": {
        "operationId": "listSpaces",
        "parameters": [
          {
            "description": "Max results (default 50, max 200)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "Max results (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Pagination cursor",
              "type": "string"
            }
          },
          {
            "description": "Search by name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "description": "Search by name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "cursor": {
                      "description": "Opaque pagination cursor",
                      "example": "eyJ0Ijoi...",
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "spaces": {
                      "items": {
                        "$ref": "#/components/schemas/Space"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "spaces",
                    "cursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Space listing"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          }
        },
        "summary": "List spaces (paginated, RLS filters by read_level)",
        "tags": [
          "Spaces"
        ],
        "x-arke-auth": "optional",
        "x-arke-related": [
          "POST /spaces",
          "GET /spaces/{id}"
        ],
        "x-arke-rules": [
          "Results filtered by your classification clearance"
        ]
      },
      "post": {
        "operationId": "createSpace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "description": {
                    "description": "Space description",
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "name": {
                    "description": "Space name",
                    "minLength": 1,
                    "type": "string"
                  },
                  "properties": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/JsonObject"
                      },
                      {
                        "description": "Arbitrary properties"
                      }
                    ]
                  },
                  "read_level": {
                    "description": "Read classification level (0-4)",
                    "maximum": 4,
                    "minimum": 0,
                    "type": "integer"
                  },
                  "write_level": {
                    "description": "Write classification level (0-4)",
                    "maximum": 4,
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "space": {
                      "$ref": "#/components/schemas/Space"
                    }
                  },
                  "required": [
                    "space"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Space created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "Create a new space",
        "tags": [
          "Spaces"
        ],
        "x-arke-auth": "required",
        "x-arke-related": [
          "GET /spaces/{id}",
          "GET /spaces"
        ],
        "x-arke-rules": [
          "You become the owner of the space"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.arke.institute"
    }
  ],
...
}